% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattanGenerator.R
\name{manhattanPlot}
\alias{manhattanPlot}
\title{Function to generate manhattan plot, QQ plot or top SNPs.}
\usage{
manhattanPlot(
  Path,
  imgName,
  title,
  outDist,
  sugg = 1e-06,
  sigg = 5e-08,
  Manhattan = T,
  pvalColName = "P_BOLT_LMM_INF",
  basePairName = "BP",
  chrName = "CHR",
  QQ = F,
  chisqCol = "CHISQ_LINREG",
  topSNPs = F,
  SNP_distance = 100
)
}
\arguments{
\item{Path}{Path to summary statistics}

\item{imgName}{Name of the images to generate.}

\item{title}{title on plots}

\item{outDist}{path to the output destination folder for plots and top snps}

\item{sugg}{suggestive threshold, default is 5e-6}

\item{sigg}{genome-wide significant threshold, default is 5e-8}

\item{Manhattan}{True/False variable for whether or not to generate the manhattan plot}

\item{pvalColName}{Name of p value column in summary statistics file}

\item{basePairName}{Name of basepair column in summary statistics file}

\item{chrName}{Name of chromosome column in summary statistics file}

\item{QQ}{True/False variable for whether or not to generate the QQ plot}

\item{chisqCol}{Name of ChiSQ column in summary statistics file}

\item{topSNPs}{True/False variable for whether or not to generate top snps files}

\item{SNP_distance}{distance used to determine independent snps}
}
\value{
Generates a manhattan plot or QQ plot. If asked to, it will also store the top snps that are either genomewide significant or suggestive to separate files.
}
\description{
Function to generate manhattan plot, QQ plot or top SNPs.
}
